ssh_agent/requirements.txt
paramiko>=3.0.0
pyyaml>=6.0
pexpect>=4.8.0
mcp>=0.1.0
